requirements.txt
google-generativeai
pandas
numpy
scipy